+++ src/components/shop/confirm/confirm-buttons.js
@@ -5,20 +5,8 @@
  */
 import { createConfirmButtons } from "../menu-buttons.js";
 import fp from "../../../../lib/lodash/fp/fp.js";
-import { buy, equip, getBalanceItem, unequip, use } from "../transact.js";
-import { collections } from "../../../core/collections.js";
-
-/* DUPES */
-const getShopConfig = scene => scene.transientData.shop.config;
-const canAffordItem = (scene, item) => item && getBalanceItem(getShopConfig(scene)).qty >= item.price;
-const itemIsInStock = (scene, item) =>
-    item && collections.get(getShopConfig(scene).shopCollections.shop).get(item.id).qty > 0;
-
-const isEquippable = item => item && item.slot;
-/* END DUPES */
-
-
-
+import { buy, equip, unequip, use } from "../transact.js";
+import { canAffordItem, isEquippable, itemIsInStock } from "./utility-rename.js";
 
 const disableActionButton = button => {
     Object.assign(button, { alpha: 0.25, tint: 0xff0000 });
@@ -27,17 +15,17 @@ const disableActionButton = button => {
 };
 
 const handleActionClick = (scene, title, action, item) => {
-    doAction({ scene, action, item });
+    actions[action]({ scene, item });
     scene._data.addedBy.scene.resume();
     scene.removeOverlay();
 };
 
-const doAction = fp.cond([
-    [args => args.action === "buy", args => buy(args.scene, args.item)],
-    [args => args.action === "equip", args => equip(args.scene, args.item)],
-    [args => args.action === "unequip", args => unequip(args.scene, args.item)],
-    [args => args.action === "use", args => use(args.scene, args.item)],
-]);
+const actions = {
+    buy: args => buy(args.scene, args.item),
+    equip: args => equip(args.scene, args.item),
+    unequip: args => unequip(args.scene, args.item),
+    use: args => use(args.scene, args.item),
+};
 
 const canBuyItem = (scene, item) => canAffordItem(scene, item) && itemIsInStock(scene, item);
