+++ test/components/shop/shop-confirm-screen.test.js
@@ -12,7 +12,7 @@ import { ShopConfirm } from "../../../src/components/shop/shop-confirm-screen.js
 import { initResizers } from "../../../src/components/shop/backgrounds.js";
 
 jest.mock("../../../src/components/shop/balance.js");
-jest.mock("../../../src/components/shop/confirm.js");
+jest.mock("../../../src/components/shop/confirm/confirm.js");
 jest.mock("../../../lib/rexuiplugin.min.js");
 jest.mock("../../../src/core/scaler.js");
